+++ R/data.R
@@ -1,0 +1,13 @@
+#' MPT dictionary
+#'
+#' The default internationalisation dictionary used by the MPT.
+#' @name mpt_dict
+#' @docType data
+NULL
+
+#' MPT item bank
+#'
+#' The MPT's item bank
+#' @name item_bank
+#' @docType data
+NULL