+++ NEWS.md
@@ -1,3 +1,8 @@
+* The package is now properly documented.
+* Change to default arguments for `mpt()`, `demo_mpt()`,
+and `standalone_mpt()`: 
+`next_item.estimator` now defaults to `'BM'`, not `'WL'`.
+
 # mpt 0.2.0
 
 * Adding French support